Remove optional AS keyword from FROM clauses
Update queries using correlation names (aliases) in FROM clause to
allow for running on Oracle, as Oracle does not support the optional
AS keyword for correlation names in the FROM keyword. See also
https://docs.oracle.com/en/database/oracle/oracle-database/19/sqlrf/Ora
cle-Compliance-To-Core-SQL2011.html#GUID-D372D906-805B-49B8-824A-D4697B
05B7F8 (E051-08, Correlation names in the FROM clause).

Using the AS keyword in query to an Oracle database gives error
ORA-00933: SQL command not properly ended.